draw support and separate winrate for P1/P2

+++ src/main/java/fr/svivien/cgbenchmark/model/request/play/PlayResponse.java
@@ -23,6 +23,6 @@ public class PlayResponse {
     public class PlayResponseSuccess {
         public List<Integer> scores;
         public List<Frame> frames;
-        public int gameId;
+        public long gameId;
     }
 }
